+++ types/dynamic/dynamic_test.go
@@ -73,13 +73,15 @@ type MyType struct {
 type MyTypeEmbedded struct {
 	Foo         string `json:"foo"`
 	Bar         int    `json:"bar"`
+	NotEmbedded Meta   `json:"not-embed"`
 	Meta        `json:"meta"`
-	NotEmbedded Meta `json:"not-embed"`
 }
 
 type Meta struct {
-	Name      string `json:"name"`
-	Namespace string `json:"namespace"`
+	Name        string            `json:"name"`
+	Namespace   string            `json:"namespace"`
+	Labels      map[string]string `json:"labels"`
+	Annotations map[string]string `json:"annotations"`
 }
 
 func (m *MyType) Get(name string) (interface{}, error) {
@@ -171,16 +173,40 @@ func TestSynthesize(t *testing.T) {
 			},
 		},
 		{
-			name:  "embedded fields",
-			input: &MyTypeEmbedded{Foo: "bar", Bar: 5, Meta: Meta{Name: "baz", Namespace: "default"}, NotEmbedded: Meta{Name: "not-baz", Namespace: "not-default"}},
+			name: "embedded fields",
+			input: &MyTypeEmbedded{
+				Foo: "bar",
+				Bar: 5,
+				Meta: Meta{
+					Name:        "baz",
+					Namespace:   "default",
+					Labels:      map[string]string{"Hi": "hello"},
+					Annotations: map[string]string{"One": "1", "Two": "2"},
+				},
+				NotEmbedded: Meta{
+					Name:        "not-baz",
+					Namespace:   "not-default",
+					Labels:      map[string]string{},
+					Annotations: map[string]string{},
+				},
+			},
 			expected: map[string]interface{}{
 				"Bar":       5,
 				"Foo":       "bar",
 				"Name":      "baz",
 				"Namespace": "default",
+				"Labels": map[string]string{
+					"Hi": "hello",
+				},
+				"Annotations": map[string]string{
+					"One": "1",
+					"Two": "2",
+				},
 				"NotEmbedded": map[string]interface{}{
-					"Name":      "not-baz",
-					"Namespace": "not-default",
+					"Name":        "not-baz",
+					"Namespace":   "not-default",
+					"Labels":      map[string]string{},
+					"Annotations": map[string]string{},
 				},
 			},
 		},